ai-service/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
numpy==1.21.2
pandas==1.3.3
opencv-python==4.5.3.56
mediapipe==0.8.9.1
tensorflow==2.7.0
torch==1.9.1
torchvision==0.10.1
scikit-learn==0.24.2
pyyaml==5.4.1
tqdm==4.62.3
gitpython==3.1.24
gunicorn==20.1.0 